Revert "move exec22 to package, call it from stage_exec (#5000)"
This reverts commit efa6dfd8ce9eb64896da6b1b4f7e7a6a5ffa09e4.
