+++ cmd/state/commands/state_recon.go
@@ -22,7 +22,7 @@ import (
 	"github.com/ledgerwatch/erigon-lib/kv"
 	kv2 "github.com/ledgerwatch/erigon-lib/kv/mdbx"
 	libstate "github.com/ledgerwatch/erigon-lib/state"
-	"github.com/ledgerwatch/erigon/cmd/state/exec22"
+	"github.com/ledgerwatch/erigon/cmd/state/state22"
 	"github.com/ledgerwatch/erigon/common"
 	"github.com/ledgerwatch/erigon/common/dbutils"
 	"github.com/ledgerwatch/erigon/consensus"
@@ -73,8 +73,8 @@ type ReconWorker struct {
 	chainConfig  *params.ChainConfig
 	logger       log.Logger
 	genesis      *core.Genesis
-	epoch        exec22.EpochReader
-	chain        exec22.ChainReader
+	epoch        state22.EpochReader
+	chain        state22.ChainReader
 	isPoSA       bool
 	posa         consensus.PoSA
 }
@@ -99,8 +99,8 @@ func NewReconWorker(lock sync.Locker, wg *sync.WaitGroup, rs *state.ReconState,
 		genesis:      genesis,
 		engine:       engine,
 	}
-	rw.epoch = exec22.NewEpochReader(chainTx)
-	rw.chain = exec22.NewChainReader(chainConfig, chainTx, blockReader)
+	rw.epoch = state22.NewEpochReader(chainTx)
+	rw.chain = state22.NewChainReader(chainConfig, chainTx, blockReader)
 	rw.posa, rw.isPoSA = engine.(consensus.PoSA)
 	return rw
 }
