+++ eth/stagedsync/stage_execute.go
@@ -32,7 +32,6 @@ import (
 	"github.com/ledgerwatch/erigon/ethdb"
 	"github.com/ledgerwatch/erigon/ethdb/olddb"
 	"github.com/ledgerwatch/erigon/ethdb/prune"
-	"github.com/ledgerwatch/erigon/node/nodecfg/datadir"
 	"github.com/ledgerwatch/erigon/params"
 	"github.com/ledgerwatch/erigon/turbo/services"
 	"github.com/ledgerwatch/erigon/turbo/shards"
@@ -59,13 +58,11 @@ type ExecuteBlockCfg struct {
 	engine        consensus.Engine
 	vmConfig      *vm.Config
 	badBlockHalt  bool
+	tmpdir        string
 	stateStream   bool
 	accumulator   *shards.Accumulator
 	blockReader   services.FullBlockReader
 	hd            *headerdownload.HeaderDownload
-
-	exec22 bool
-	dirs   datadir.Dirs
 }
 
 func StageExecuteBlocksCfg(
@@ -79,7 +76,7 @@ func StageExecuteBlocksCfg(
 	accumulator *shards.Accumulator,
 	stateStream bool,
 	badBlockHalt bool,
-	dirs datadir.Dirs,
+	tmpdir string,
 	blockReader services.FullBlockReader,
 	hd *headerdownload.HeaderDownload,
 ) ExecuteBlockCfg {
@@ -91,7 +88,7 @@ func StageExecuteBlocksCfg(
 		chainConfig:   chainConfig,
 		engine:        engine,
 		vmConfig:      vmConfig,
-		dirs:          dirs,
+		tmpdir:        tmpdir,
 		accumulator:   accumulator,
 		stateStream:   stateStream,
 		badBlockHalt:  badBlockHalt,
@@ -207,33 +204,7 @@ func newStateReaderWriter(
 	return stateReader, stateWriter, nil
 }
 
-func executeBlocks22(s *StageState, u Unwinder, tx kv.RwTx, toBlock uint64, ctx context.Context, cfg ExecuteBlockCfg, initialCycle bool) (err error) {
-	/*
-		//TODO: create agg22 right here
-		var agg22DB *kv.RoDB
-		//TODO: fill txNums
-		var txNums []uint64
-		//TODO: create agg
-		var agg *state2.Aggregator22
-		//TODO: create rs
-
-		reconDbPath := path.Join(cfg.dirs.DataDir, "db22")
-		if reset && dir.Exist(reconDbPath) {
-			if err = os.RemoveAll(reconDbPath); err != nil {
-				return err
-			}
-		}
-		dir.MustExist(reconDbPath)
-
-		return exec22.Exec22(ctx, s, s.BlockNumber+1, 1, agg22DB, cfg.db, tx, rs, cfg.blockReader, cfg.snapshots, log.New(), cfg.engine, toBlock, cfg.chainConfig, cfg.genesis, initialCycle)
-	*/
-	return nil
-}
 func SpawnExecuteBlocksStage(s *StageState, u Unwinder, tx kv.RwTx, toBlock uint64, ctx context.Context, cfg ExecuteBlockCfg, initialCycle bool) (err error) {
-	if cfg.exec22 {
-		return executeBlocks22(s, u, tx, toBlock, ctx, cfg, initialCycle)
-	}
-
 	quit := ctx.Done()
 	useExternalTx := tx != nil
 	if !useExternalTx {
@@ -270,7 +241,7 @@ func SpawnExecuteBlocksStage(s *StageState, u Unwinder, tx kv.RwTx, toBlock uint
 
 	var batch ethdb.DbWithPendingMutations
 	// state is stored through ethdb batches
-	batch = olddb.NewHashBatch(tx, quit, cfg.dirs.Tmp)
+	batch = olddb.NewHashBatch(tx, quit, cfg.tmpdir)
 
 	defer batch.Rollback()
 	// changes are stored through memory buffer
@@ -367,7 +338,7 @@ Loop:
 				// TODO: This creates stacked up deferrals
 				defer tx.Rollback()
 			}
-			batch = olddb.NewHashBatch(tx, quit, cfg.dirs.Tmp)
+			batch = olddb.NewHashBatch(tx, quit, cfg.tmpdir)
 			// TODO: This creates stacked up deferrals
 			defer batch.Rollback()
 		}
@@ -491,7 +462,7 @@ func unwindExecutionStage(u *UnwindState, s *StageState, tx kv.RwTx, quit <-chan
 		accumulator.StartChange(u.UnwindPoint, hash, txs, true)
 	}
 
-	changes := etl.NewCollector(logPrefix, cfg.dirs.Tmp, etl.NewOldestEntryBuffer(etl.BufferOptimalSize))
+	changes := etl.NewCollector(logPrefix, cfg.tmpdir, etl.NewOldestEntryBuffer(etl.BufferOptimalSize))
 	defer changes.Close()
 	errRewind := changeset.RewindData(tx, s.BlockNumber, u.UnwindPoint, changes, quit)
 	if errRewind != nil {
